fix : Make admin access dashboard only using middleware

+++ routes/web.php
@@ -10,21 +10,21 @@ use App\Http\Controllers\ProductController;
 
 use Illuminate\Support\Facades\Route;
 
-Route::get('/', [CategoryController::class,'index'])->name('homePage');
-Route::get('/shop/{category}',[ProductController::class,'index'])->where('category','[0-9]+')->name('shop');
-Route::get('/site/product/{id}',[ProductController::class,'show'])->where('id','[0-9]+')->name('site.product.show');
-Route::get('/shop',[ProductController::class,'getAllProducts'])->name('products');
-Route::get('/cart',[CardController::class,'index'])->name('cart');
-Route::get('/card/add/{id}',[CardController::class,'add'])->where('id','[0-9]+')->name('addToCard');
-Route::delete('/card/{id}',[CardController::class,'delete'])->where('id','[0-9]+')->name('deleteFromCard');
-Route::get('/checkout',[OrderController::class,'index'])->name('checkout');
-Route::post('/checkout/store',[OrderController::class,'store'])->name('checkout.store');
+Route::get('/', [CategoryController::class,'index'])->name('homePage')->middleware(['auth','roleCheck']);
+Route::get('/shop/{category}',[ProductController::class,'index'])->where('category','[0-9]+')->name('shop')->middleware(['auth','roleCheck']);
+Route::get('/site/product/{id}',[ProductController::class,'show'])->where('id','[0-9]+')->name('site.product.show')->middleware(['auth','roleCheck']);
+Route::get('/shop',[ProductController::class,'getAllProducts'])->name('products')->middleware(['auth','roleCheck']);
+Route::get('/cart',[CardController::class,'index'])->name('cart')->middleware(['auth','roleCheck']);
+Route::get('/card/add/{id}',[CardController::class,'add'])->where('id','[0-9]+')->name('addToCard')->middleware(['auth','roleCheck']);
+Route::delete('/card/{id}',[CardController::class,'delete'])->where('id','[0-9]+')->name('deleteFromCard')->middleware(['auth','roleCheck']);
+Route::get('/checkout',[OrderController::class,'index'])->name('checkout')->middleware(['auth','roleCheck']);
+Route::post('/checkout/store',[OrderController::class,'store'])->name('checkout.store')->middleware(['auth','roleCheck']);
 
 
 //Dashboard routes
-Route::resource('/product',DashboardProductController::class)->middleware(['auth','roleCheck']);
-Route::resource('/category',DashboardCategoryController::class)->middleware(['auth','roleCheck']);
-Route::resource('/order',DashboardOrderController::class)->middleware(['auth','roleCheck']);
+Route::resource('/product',DashboardProductController::class)->middleware(['auth']);
+Route::resource('/category',DashboardCategoryController::class)->middleware(['auth']);
+Route::resource('/order',DashboardOrderController::class)->middleware(['auth']);
 
 
 Auth::routes();
